Ensure the correct go is found
gopls relies on PATH to find the go executable. This sets the path if necessary to ensure the go version specified in preferences is found.

This also changes watchConfigVarCb() to only call the callback if the effective value of the preference has changed. This avoids restart storms when adjusting preferences as Nova tends to send out redundant notifications.